requirements.txt
asgiref==3.5.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==3.2.1
django-cors-headers==3.11.0
django-dotenv==1.4.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
gunicorn==21.2.0
idna==3.4
keras==2.13.1
numpy==1.24.3
packaging==23.2
Pillow==9.5.0
psycopg2==2.9.9
PyJWT==2.3.0
pyproj==3.5.0
pytz==2021.3
requests==2.31.0
shapely==2.0.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==2.0.3
whitenoise==6.6.0
xlwt==1.3.0
